Create README.md
I was having trouble understanding the insert function with heaps. From the notes I understand that first you must insert the value at the end of the heap and then compare it to the parent. Then if the parent is less than the value you swap the values. Otherwise you are done. I am using an array implementation. This is just the header file.